// src/types/payload.ts
export interface RemoteFileUpload {
  id: number;
  /**
   * Type of the remote file
   */
  type: 'image' | 'document';
  /**
   * The service provider this file is associated with
   */
  serviceProvider?: (number | null) | ServiceProvider;
  /**
   * Remote file URL from our remote storage
   */
  remoteURL: string;
  /**
   * Optional description of the file
   */
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}